fix(step3/Car,GameRule): 패키지 구조 변경
