--- src/main/java/racingcar/GameRule.java
@@ -1,9 +1,0 @@
-package racingcar;
-
-public class GameRule {
-	private GameRule() {}
-
-	public static boolean isMoveCar(int randomValue) {
-		return randomValue >= 4;
-	}
-}
