+++ src/main/java/racingcar/domain/Car.java
@@ -1,0 +1,12 @@
+package racingcar.domain;
+
+public class Car {
+	private int move = 0;
+
+	public int run(int randomNumber) {
+		if (GameRule.isMoveCar(randomNumber)) {
+			move++;
+		}
+		return move;
+	}
+}
