+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,15 +1,14 @@
-package racingcar;
+package racingcar.domain;
 
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import exception.OutOfRangeException;
+import racingcar.domain.Car;
 
 public class CarTest {
